+++ Toggly.FeatureManagement.Nuxt/nuxt-toggly-core/src/constants.ts
@@ -6,7 +6,6 @@ export const DEFAULT_CONFIG = {
   environment: 'Production',
   refreshInterval: 180000, // 3 minutes
   showFeatureDuringEvaluation: false,
-  enableLiveUpdates: false,
 } as const
 
 /**