+++ Toggly.FeatureManagement.Nuxt/nuxt-toggly-core/src/index.ts
@@ -29,7 +29,22 @@ export {
   isPlainObject,
   debounce,
   createDeferred,
+  isBrowser,
+  isServer,
+  isEdgeRuntime,
 } from './utils'
 
 // Client
 export { createTogglyClient } from './client'
+
+// Live socket helpers
+export {
+  resolveWebSocketConstructor,
+  openLiveSocket,
+  dispatchLiveMessage,
+} from './live-socket'
+export type {
+  WebSocketConstructor,
+  LiveSocket,
+  LiveSocketHandlers,
+} from './live-socket'